+++ src/parsing/parsenode.ts
@@ -358,6 +358,7 @@ export interface Comparator extends BaseComparator {
 interface DoubleInequality extends Base {
   // Not sure how to get DoubleInequality
   type: "DoubleInequality";
+  args: [ChildExprNode, string, ChildExprNode, string, ChildExprNode];
   _operators: [Expression, Expression];
   _expressions: [Expression, Expression];
   _indicator: unknown;
@@ -618,7 +619,8 @@ export type ChildExprNode =
   | Assignment
   // Seed + ExtendSeed only used in SeededFunctionCalls?
   | Seed
-  | ExtendSeed;
+  | ExtendSeed
+  | DoubleInequality;
 
 // These can only occur after further transformation or something
 type IrrelevantExprNode =